Move StateTrackers to their own package, and unit mods out of CO-mod-ville
