+++ src/CommandingOfficers/Tech.java
@@ -11,9 +11,6 @@ import java.util.Set;
 
 import AI.AICombatUtils;
 import AI.AIUtils;
-import CommandingOfficers.Modifiers.UnitDamageModifier;
-import CommandingOfficers.Modifiers.UnitDefenseModifier;
-import CommandingOfficers.Modifiers.UnitTypeFilter;
 import Engine.GameAction.UnitSpawnAction;
 import Engine.GameScenario;
 import Engine.UnitActionFactory;
@@ -22,7 +19,10 @@ import Engine.XYCoord;
 import Engine.Combat.DamagePopup;
 import Engine.GameEvents.CommanderDefeatEvent;
 import Engine.GameEvents.GameEventQueue;
+import Engine.UnitMods.UnitDamageModifier;
+import Engine.UnitMods.UnitDefenseModifier;
 import Engine.UnitMods.UnitModifier;
+import Engine.UnitMods.UnitTypeFilter;
 import Engine.GameEvents.CreateUnitEvent;
 import Terrain.GameMap;
 import Terrain.MapLocation;
