+++ src/Engine/GameInstance.java
@@ -15,8 +15,8 @@ import Engine.GameEvents.GameEventListener;
 import Engine.GameEvents.GameEventQueue;
 import Engine.GameEvents.MapChangeEvent;
 import Engine.GameEvents.TurnInitEvent;
+import Engine.StateTrackers.StateTracker;
 import Engine.UnitMods.SandstormModifier;
-import Engine.UnitMods.StateTracker;
 import Terrain.Environment;
 import Terrain.Environment.Weathers;
 import Terrain.MapLocation;
