+++ src/Engine/UnitMods/SandstormModifier.java
@@ -1,6 +1,5 @@
 package Engine.UnitMods;
 
-import CommandingOfficers.Modifiers.UnitIndirectRangeModifier;
 import Terrain.Environment.Weathers;
 import Units.UnitContext;
