+++ src/Test/TestCombatMods.java
@@ -3,7 +3,6 @@ package Test;
 import CommandingOfficers.Cinder;
 import CommandingOfficers.Commander;
 import CommandingOfficers.Venge;
-import CommandingOfficers.Modifiers.UnitDamageModifier;
 import Engine.GameInstance;
 import Engine.GameScenario;
 import Engine.Utils;
@@ -11,8 +10,9 @@ import Engine.Combat.BattleSummary;
 import Engine.Combat.CombatEngine;
 import Engine.GameEvents.GameEventListener;
 import Engine.GameEvents.ResupplyEvent;
+import Engine.StateTrackers.DamageDealtToIncomeConverter;
 import Engine.UnitActionLifecycles.BattleLifecycle;
-import Engine.UnitMods.DamageDealtToIncomeConverter;
+import Engine.UnitMods.UnitDamageModifier;
 import Terrain.MapLibrary;
 import Terrain.MapMaster;
 import Units.Unit;
