+++ src/UI/Art/SpriteArtist/UnitSpriteSet.java
@@ -11,7 +11,7 @@ import java.util.*;
 
 import CommandingOfficers.Commander;
 import Engine.GameInstance;
-import Engine.UnitMods.StateTracker;
+import Engine.StateTrackers.StateTracker;
 import UI.UIUtils;
 import UI.UIUtils.Faction;
 import UI.UnitMarker;
